+++ presentation/src/main/java/com/gta/presentation/ui/cardetail/CarDetailViewModel.kt
@@ -8,11 +8,7 @@ import com.gta.domain.model.CarState
 import com.gta.domain.usecase.cardetail.GetCarDetailDataUseCase
 import com.gta.domain.usecase.cardetail.IsMyCarUseCase
 import com.gta.domain.usecase.cardetail.IsNowMyRentCarUseCase
-import com.gta.presentation.model.* // ktlint-disable no-wildcard-imports
-import com.gta.presentation.model.carDetail.BtnType
-import com.gta.presentation.model.carDetail.CarInfo
-import com.gta.presentation.model.carDetail.PriceType
-import com.gta.presentation.model.carDetail.toCarInfo
+import com.gta.presentation.model.carDetail.* // ktlint-disable no-wildcard-imports
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.* // ktlint-disable no-wildcard-imports
 import kotlinx.coroutines.launch
