+++ presentation/src/main/res/layout/include_owner_profile.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="owner"
-            type="com.gta.presentation.model.CarOwner" />
+            type="com.gta.presentation.model.carDetail.CarOwner" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
